feat: refactor the folder structure and link
