--- app/(app)/profile/index.tsx
@@ -1,3 +1,0 @@
-import Profile from '@/screens/profile';
-
-export default Profile;
