--- app/(app)/user/[id].tsx
@@ -1,3 +1,0 @@
-import User from '@/screens/user';
-
-export default User;
