+++ src/app/(app)/(tabs)/(drawer)/feeds.tsx
@@ -1,4 +1,4 @@
-import { Text, View } from '@/components/Themed';
+import { Text, View } from '@/src/components/Themed';
 import { Link } from 'expo-router';
 import React from 'react';
 import { Pressable } from 'react-native';
