+++ src/app/(app)/modal.tsx
@@ -1,8 +1,8 @@
 import { StatusBar } from 'expo-status-bar';
 import { Button, Platform, StyleSheet } from 'react-native';
 
-import EditScreenInfo from '@/components/EditScreenInfo';
-import { Text, View } from '@/components/Themed';
+import EditScreenInfo from '@/src/components/EditScreenInfo';
+import { Text, View } from '@/src/components/Themed';
 import { router } from 'expo-router';
 
 export default function ModalScreen() {
