+++ src/app/(app)/profile/index.tsx
@@ -1,0 +1,3 @@
+import Profile from '@/src/screens/profile';
+
+export default Profile;
