+++ src/app/(app)/user/[id].tsx
@@ -1,0 +1,3 @@
+import User from '@/src/screens/user';
+
+export default User;
