+++ src/app/+not-found.tsx
@@ -1,7 +1,7 @@
 import { Link, Stack } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { Text, View } from '@/components/Themed';
+import { Text, View } from '@/src/components/Themed';
 import React from 'react';
 
 export default function NotFoundScreen() {
