+++ src/app/_layout.tsx
@@ -10,8 +10,8 @@ import * as SplashScreen from 'expo-splash-screen';
 import { useEffect } from 'react';
 import 'react-native-reanimated';
 
-import { useColorScheme } from '@/components/useColorScheme';
-import { AuthProvider } from '@/context/AuthProvider';
+import { useColorScheme } from '@/src/components/useColorScheme';
+import { AuthProvider } from '@/src/context/AuthProvider';
 
 export {
   // Catch any errors thrown by the Layout component.
@@ -28,7 +28,7 @@ SplashScreen.preventAutoHideAsync();
 
 export default function RootLayout() {
   const [loaded, error] = useFonts({
-    SpaceMono: require('../assets/fonts/SpaceMono-Regular.ttf'),
+    SpaceMono: require('../../assets/fonts/SpaceMono-Regular.ttf'),
     ...FontAwesome.font,
   });
